requirements.txt
pygame>=2.5.0
numpy>=1.24.0
networkx>=3.0
openai>=1.0.0
google-generativeai>=0.3.0
